Move _socket.ssl to _ssl._SSLSocket
