+++ Src/IronPython.Modules/_ssl.cs
@@ -9,19 +9,21 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Linq;
+using System.Net.Security;
+using System.Net.Sockets;
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
+using System.Security.Authentication;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
 using System.Text;
-
-using Microsoft.Scripting.Runtime;
-using Microsoft.Scripting.Utils;
 
 using IronPython.Runtime;
 using IronPython.Runtime.Exceptions;
 using IronPython.Runtime.Operations;
 using IronPython.Runtime.Types;
+
+using Microsoft.Scripting.Runtime;
+using Microsoft.Scripting.Utils;
 
 [assembly: PythonModule("_ssl", typeof(IronPython.Modules.PythonSsl))]
 namespace IronPython.Modules {
@@ -105,7 +107,6 @@ namespace IronPython.Modules {
 
         #endregion
 
-        #region SSLContext
         [PythonType]
         public class _SSLContext {
             private X509Certificate2Collection _cert_store = new X509Certificate2Collection();
@@ -197,15 +198,409 @@ namespace IronPython.Modules {
             }
 
             public object _wrap_socket(CodeContext context, PythonSocket.socket sock = null, bool server_side = false, string server_hostname = null, object ssl_sock = null) {
-                return new PythonSocket.ssl(context, sock, server_side, null, _cafile, verify_mode, protocol | options, null, _cert_store) { _serverHostName = server_hostname };
+                return new _SSLSocket(context, sock, server_side, null, _cafile, verify_mode, protocol | options, null, _cert_store) { _serverHostName = server_hostname };
             }
         }
 
-        #endregion
+        [PythonType]
+        public class _SSLSocket {
+            private SslStream _sslStream;
+            private PythonSocket.socket _socket;
+            private readonly X509Certificate2Collection _certCollection;
+            private readonly X509Certificate _cert;
+            private readonly int _protocol, _certsMode;
+            private readonly bool _validate, _serverSide;
+            private readonly CodeContext _context;
+            private readonly RemoteCertificateValidationCallback _callback;
+            private Exception _validationFailure;
+            internal string _serverHostName;
 
-        public static PythonType SSLType = DynamicHelpers.GetPythonTypeFromType(typeof(PythonSocket.ssl));
+            public _SSLSocket(CodeContext context, PythonSocket.socket sock, string keyfile = null, string certfile = null, X509Certificate2Collection certs = null) {
+                _context = context;
+                _sslStream = new SslStream(new NetworkStream(sock._socket, false), true, CertValidationCallback);
+                _socket = sock;
+                _protocol = PythonSsl.PROTOCOL_SSLv23 | PythonSsl.OP_NO_SSLv2 | PythonSsl.OP_NO_SSLv3;
+                _validate = false;
+                _certCollection = certs ?? new X509Certificate2Collection();
+            }
 
-        public static PythonSocket.ssl sslwrap(
+            internal _SSLSocket(CodeContext context,
+               PythonSocket.socket sock,
+               bool server_side,
+               string keyfile = null,
+               string certfile = null,
+               int certs_mode = PythonSsl.CERT_NONE,
+               int protocol = (PythonSsl.PROTOCOL_SSLv23 | PythonSsl.OP_NO_SSLv2 | PythonSsl.OP_NO_SSLv3),
+               string cacertsfile = null,
+               X509Certificate2Collection certs = null) {
+                if (sock == null) {
+                    throw PythonOps.TypeError("expected socket object, got None");
+                }
+
+                _serverSide = server_side;
+                bool validate;
+                _certsMode = certs_mode;
+
+                RemoteCertificateValidationCallback callback;
+                switch (certs_mode) {
+                    case PythonSsl.CERT_NONE:
+                        validate = false;
+                        callback = CertValidationCallback;
+                        break;
+                    case PythonSsl.CERT_OPTIONAL:
+                        validate = true;
+                        callback = CertValidationCallbackOptional;
+                        break;
+                    case PythonSsl.CERT_REQUIRED:
+                        validate = true;
+                        callback = CertValidationCallbackRequired;
+                        break;
+                    default:
+                        throw new InvalidOperationException(String.Format("bad certs_mode: {0}", certs_mode));
+                }
+
+                _callback = callback;
+
+                if (certs != null) {
+                    _certCollection = certs;
+                }
+
+                if (certfile != null) {
+                    _cert = PythonSsl.ReadCertificate(context, certfile);
+                }
+
+                if (cacertsfile != null) {
+                    _certCollection = new X509Certificate2Collection(new[] { PythonSsl.ReadCertificate(context, cacertsfile) });
+                }
+
+                _socket = sock;
+
+                EnsureSslStream(false);
+
+                _protocol = protocol;
+                _validate = validate;
+                _context = context;
+            }
+
+            private void EnsureSslStream(bool throwWhenNotConnected) {
+                if (_sslStream == null && _socket._socket.Connected) {
+                    if (_serverSide) {
+                        _sslStream = new SslStream(
+                            new NetworkStream(_socket._socket, false),
+                            true,
+                            _callback
+                        );
+                    } else {
+                        _sslStream = new SslStream(
+                            new NetworkStream(_socket._socket, false),
+                            true,
+                            _callback,
+                            CertSelectLocal
+                        );
+                    }
+                }
+
+                if (throwWhenNotConnected && _sslStream == null) {
+                    throw PythonExceptions.CreateThrowable(PythonSocket.error, 10057, "A request to send or receive data was disallowed because the socket is not connected.");
+                }
+            }
+
+            internal bool CertValidationCallback(object sender, X509Certificate certificate, X509Chain chain, SslPolicyErrors sslPolicyErrors) {
+                return true;
+            }
+
+            internal bool CertValidationCallbackOptional(object sender, X509Certificate certificate, X509Chain chain, SslPolicyErrors sslPolicyErrors) {
+                if (!_serverSide) {
+                    if (certificate != null && sslPolicyErrors != SslPolicyErrors.None) {
+                        ValidateCertificate(certificate, chain, sslPolicyErrors);
+                    }
+                }
+
+                return true;
+            }
+
+            internal X509Certificate CertSelectLocal(object sender, string targetHost, X509CertificateCollection collection, X509Certificate remoteCertificate, string[] acceptableIssuers) {
+                if (acceptableIssuers != null && acceptableIssuers.Length > 0 && collection != null && collection.Count > 0) {
+                    // Use the first certificate that is from an acceptable issuer.
+                    foreach (X509Certificate certificate in collection) {
+                        string issuer = certificate.Issuer;
+                        if (Array.IndexOf(acceptableIssuers, issuer) != -1)
+                            return certificate;
+                    }
+                }
+
+                if (collection != null && collection.Count > 0) {
+                    return collection[0];
+                }
+
+                return null;
+            }
+
+            internal bool CertValidationCallbackRequired(object sender, X509Certificate certificate, X509Chain chain, SslPolicyErrors sslPolicyErrors) {
+                if (!_serverSide) {
+                    // client check
+                    if (certificate == null) {
+                        ValidationError(SslPolicyErrors.None);
+                    } else if (sslPolicyErrors != SslPolicyErrors.None) {
+                        ValidateCertificate(certificate, chain, sslPolicyErrors);
+                    }
+                }
+
+                return true;
+            }
+
+            private void ValidateCertificate(X509Certificate certificate, X509Chain chain, SslPolicyErrors sslPolicyErrors) {
+                chain = new X509Chain();
+                X509Certificate2Collection certificates = new X509Certificate2Collection();
+                foreach (object cert in _certCollection) {
+                    if (cert is X509Certificate2) {
+                        certificates.Add((X509Certificate2)cert);
+                    } else if (cert is X509Certificate) {
+                        certificates.Add(new X509Certificate2((X509Certificate)cert));
+                    }
+                }
+                chain.ChainPolicy.ExtraStore.AddRange(certificates);
+                chain.Build(new X509Certificate2(certificate));
+
+                if (chain.ChainStatus.Length > 0) {
+                    foreach (var elem in chain.ChainStatus) {
+                        if (elem.Status == X509ChainStatusFlags.UntrustedRoot) {
+                            bool isOk = false;
+                            foreach (var cert in _certCollection) {
+                                if (certificate.Issuer == cert.Subject) {
+                                    isOk = true;
+                                }
+                            }
+
+                            if (isOk) {
+                                continue;
+                            }
+                        }
+
+                        ValidationError(sslPolicyErrors);
+                        break;
+                    }
+                }
+            }
+
+            private void ValidationError(object reason) {
+                _validationFailure = PythonExceptions.CreateThrowable(PythonSsl.SSLError(_context), "errors while validating certificate chain: ", reason.ToString());
+            }
+
+            public void do_handshake() {
+                try {
+                    // make sure the remote side hasn't shutdown before authenticating so we don't
+                    // hang if we're in blocking mode.
+#pragma warning disable 219 // unused variable
+                    int available = _socket._socket.Available;
+#pragma warning restore 219
+                } catch (SocketException) {
+                    throw PythonExceptions.CreateThrowable(PythonExceptions.OSError, "socket closed before handshake");
+                }
+
+                EnsureSslStream(true);
+
+                var enabledSslProtocols = GetProtocolType(_protocol);
+
+                try {
+                    if (_serverSide) {
+                        _sslStream.AuthenticateAsServer(_cert, _certsMode == PythonSsl.CERT_REQUIRED, enabledSslProtocols, false);
+                    } else {
+
+                        var collection = new X509CertificateCollection();
+
+                        if (_cert != null) {
+                            collection.Add(_cert);
+                        }
+                        _sslStream.AuthenticateAsClient(_serverHostName ?? _socket._hostName, collection, enabledSslProtocols, false);
+                    }
+                } catch (AuthenticationException e) {
+                    ((IDisposable)_socket._socket).Dispose();
+                    throw PythonExceptions.CreateThrowable(PythonSsl.SSLError(_context), "errors while performing handshake: ", e.ToString());
+                }
+
+                if (_validationFailure != null) {
+                    throw _validationFailure;
+                }
+            }
+
+            public PythonSocket.socket shutdown() {
+                _sslStream.Dispose();
+                return _socket;
+            }
+
+            /* supported communication based upon what the client & server specify
+             * as per the CPython docs:
+             * client / server SSLv2 SSLv3 SSLv23 TLSv1 TLSv1.1 TLSv1.2
+                         SSLv2   yes    no    yes    no      no      no
+                         SSLv3    no   yes    yes    no      no      no
+                        SSLv23    no   yes    yes   yes     yes     yes
+                         TLSv1    no    no    yes   yes      no      no
+                       TLSv1.1    no    no    yes    no     yes      no
+                       TLSv1.2    no    no    yes    no      no     yes
+             */
+
+            private static SslProtocols GetProtocolType(int type) {
+                SslProtocols result = SslProtocols.None;
+
+                switch (type & ~PythonSsl.OP_NO_ALL) {
+#pragma warning disable CA5397 // Do not use deprecated SslProtocols values
+#pragma warning disable CS0618 // Type or member is obsolete
+                    case PythonSsl.PROTOCOL_SSLv2:
+                        result = SslProtocols.Ssl2;
+                        break;
+                    case PythonSsl.PROTOCOL_SSLv3:
+                        result = SslProtocols.Ssl3;
+                        break;
+                    case PythonSsl.PROTOCOL_SSLv23:
+                        result = SslProtocols.Ssl2 | SslProtocols.Ssl3 | SslProtocols.Tls | SslProtocols.Tls11 | SslProtocols.Tls12;
+                        break;
+#pragma warning restore CS0618 // Type or member is obsolete
+                    case PythonSsl.PROTOCOL_TLSv1:
+                        result = SslProtocols.Tls;
+                        break;
+                    case PythonSsl.PROTOCOL_TLSv1_1:
+                        result = SslProtocols.Tls11;
+                        break;
+#pragma warning restore CA5397 // Do not use deprecated SslProtocols values
+                    case PythonSsl.PROTOCOL_TLSv1_2:
+                        result = SslProtocols.Tls12;
+                        break;
+                    default:
+                        throw new InvalidOperationException("bad ssl protocol type: " + type);
+                }
+                // Filter out requested protocol exclusions:
+#pragma warning disable CA5397 // Do not use deprecated SslProtocols values
+#pragma warning disable CS0618 // Type or member is obsolete
+                result &= (type & PythonSsl.OP_NO_SSLv3) != 0 ? ~SslProtocols.Ssl3 : ~SslProtocols.None;
+                result &= (type & PythonSsl.OP_NO_SSLv2) != 0 ? ~SslProtocols.Ssl2 : ~SslProtocols.None;
+#pragma warning restore CS0618 // Type or member is obsolete
+                result &= (type & PythonSsl.OP_NO_TLSv1) != 0 ? ~SslProtocols.Tls : ~SslProtocols.None;
+                result &= (type & PythonSsl.OP_NO_TLSv1_1) != 0 ? ~SslProtocols.Tls11 : ~SslProtocols.None;
+                result &= (type & PythonSsl.OP_NO_TLSv1_2) != 0 ? ~SslProtocols.Tls12 : ~SslProtocols.None;
+#pragma warning restore CA5397 // Do not use deprecated SslProtocols values
+                return result;
+            }
+
+            public PythonTuple cipher() {
+                if (_sslStream != null && _sslStream.IsAuthenticated) {
+                    return PythonTuple.MakeTuple(
+                        _sslStream.CipherAlgorithm.ToString(),
+                        ProtocolToPython(),
+                        _sslStream.CipherStrength
+                    );
+                }
+                return null;
+            }
+
+            private string ProtocolToPython() {
+                switch (_sslStream.SslProtocol) {
+#pragma warning disable CA5397 // Do not use deprecated SslProtocols values
+#pragma warning disable CS0618 // Type or member is obsolete
+                    case SslProtocols.Ssl2: return "SSLv2";
+                    case SslProtocols.Ssl3: return "TLSv1/SSLv3";
+#pragma warning restore CS0618 // Type or member is obsolete
+                    case SslProtocols.Tls: return "TLSv1";
+#pragma warning restore CA5397 // Do not use deprecated SslProtocols values
+                    default: return _sslStream.SslProtocol.ToString();
+                }
+            }
+
+            public object peer_certificate(bool binary_form) {
+                var peerCert = _sslStream?.RemoteCertificate;
+
+                if (peerCert != null) {
+                    if (binary_form) {
+                        return peerCert.GetRawCertData().MakeString();
+                    } else if (_validate) {
+                        return CertificateToPython(_context, peerCert);
+                    }
+                }
+                return null;
+            }
+
+            public int pending() {
+                return _socket._socket.Available;
+            }
+
+            [Documentation("issuer() -> issuer_certificate\n\n"
+                + "Returns a string that describes the issuer of the server's certificate. Only useful for debugging purposes."
+                )]
+            public string issuer() {
+                if (_sslStream != null && _sslStream.IsAuthenticated) {
+                    X509Certificate remoteCertificate = _sslStream.RemoteCertificate;
+                    if (remoteCertificate != null) {
+                        return remoteCertificate.Issuer;
+                    } else {
+                        return String.Empty;
+                    }
+                }
+                return String.Empty;
+            }
+
+            [Documentation(@"read([len]) -> string
+
+Read up to len bytes from the SSL socket.")]
+            public object read(CodeContext/*!*/ context, int len, ByteArray buffer = null) {
+                EnsureSslStream(true);
+
+                try {
+                    byte[] buf = new byte[2048];
+                    MemoryStream result = new MemoryStream(len);
+                    while (true) {
+                        int readLength = (len < buf.Length) ? len : buf.Length;
+                        int bytes = _sslStream.Read(buf, 0, readLength);
+                        if (bytes > 0) {
+                            result.Write(buf, 0, bytes);
+                            len -= bytes;
+                        }
+
+                        if (bytes == 0 || len == 0 || bytes < readLength) {
+                            var res = result.ToArray();
+                            if (buffer == null)
+                                return Bytes.Make(res);
+
+                            // TODO: get rid of the MemoryStream and write directly to the buffer
+                            buffer[new Slice(0, res.Length)] = res;
+                            return res.Length;
+                        }
+                    }
+                } catch (Exception e) {
+                    throw PythonSocket.MakeException(context, e);
+                }
+            }
+
+            [Documentation("server() -> server_certificate\n\n"
+                + "Returns a string that describes the server's certificate. Only useful for debugging purposes."
+                )]
+            public string server() {
+                if (_sslStream != null && _sslStream.IsAuthenticated) {
+                    X509Certificate remoteCertificate = _sslStream.RemoteCertificate;
+                    if (remoteCertificate != null) {
+                        return remoteCertificate.Subject;
+                    }
+                }
+                return String.Empty;
+            }
+
+            [Documentation(@"write(s) -> len
+
+Writes the string s into the SSL object.  Returns the number
+of bytes written.")]
+            public int write(CodeContext/*!*/ context, Bytes data) {
+                EnsureSslStream(true);
+
+                byte[] buffer = data.UnsafeByteArray;
+                try {
+                    _sslStream.Write(buffer);
+                    return buffer.Length;
+                } catch (Exception e) {
+                    throw PythonSocket.MakeException(context, e);
+                }
+            }
+        }
+
+        public static _SSLSocket sslwrap(
             CodeContext context,
             PythonSocket.socket socket,
             bool server_side,
@@ -215,7 +610,7 @@ namespace IronPython.Modules {
             int protocol = (PythonSsl.PROTOCOL_SSLv23 | PythonSsl.OP_NO_SSLv2 | PythonSsl.OP_NO_SSLv3),
             string cacertsfile = null,
             object ciphers = null) {
-            return new PythonSocket.ssl(
+            return new _SSLSocket(
                 context,
                 socket,
                 server_side,
@@ -327,13 +722,13 @@ namespace IronPython.Modules {
             return CertificateToPython(context, cert);
         }
 
-        internal static PythonDictionary CertificateToPython(CodeContext context, X509Certificate cert) {
+        private static PythonDictionary CertificateToPython(CodeContext context, X509Certificate cert) {
             if (cert is X509Certificate2 cert2)
                 return CertificateToPython(context, cert2);
             return CertificateToPython(context, new X509Certificate2(cert.GetRawCertData()));
         }
 
-        internal static PythonDictionary CertificateToPython(CodeContext context, X509Certificate2 cert) {
+        private static PythonDictionary CertificateToPython(CodeContext context, X509Certificate2 cert) {
             var dict = new CommonDictionaryStorage();
 
             dict.AddNoLock("notAfter", ToPythonDateFormat(cert.NotAfter));
@@ -454,9 +849,7 @@ namespace IronPython.Modules {
             return null;
         }
 
-
-        internal static X509Certificate2 ReadCertificate(CodeContext context, string filename) {
-
+        private static X509Certificate2 ReadCertificate(CodeContext context, string filename) {
             string[] lines;
             try {
                 lines = File.ReadAllLines(filename);
